Release Unified VPN 0.0.6

+++ androidApp/src/main/jni/hev-socks5-tunnel/src/core/src/hev-rbtree.h
@@ -14,7 +14,6 @@
 #define __HEV_RBTREE_H__
 
 #include <stddef.h>
-#include <stdint.h>
 
 typedef struct _HevRBTree HevRBTree;
 typedef struct _HevRBTreeNode HevRBTreeNode;
@@ -26,28 +25,28 @@ struct _HevRBTree
 
 struct _HevRBTreeNode
 {
-    uintptr_t __parent_color;
+    unsigned long __parent_color;
     HevRBTreeNode *right;
     HevRBTreeNode *left;
-} __attribute__ ((aligned (sizeof (uintptr_t))));
+} __attribute__ ((aligned (sizeof (long))));
 
 static inline int
 hev_rbtree_node_empty (HevRBTreeNode *node)
 {
-    return node->__parent_color == (uintptr_t)node;
+    return node->__parent_color == (unsigned long)node;
 }
 
 static inline HevRBTreeNode *
 hev_rbtree_node_parent (HevRBTreeNode *node)
 {
-    return (HevRBTreeNode *)(node->__parent_color & ~(uintptr_t)3);
+    return (HevRBTreeNode *)(node->__parent_color & ~3);
 }
 
 static inline void
 hev_rbtree_node_link (HevRBTreeNode *node, HevRBTreeNode *parent,
                       HevRBTreeNode **link)
 {
-    node->__parent_color = (uintptr_t)parent;
+    node->__parent_color = (unsigned long)parent;
     node->left = node->right = NULL;
 
     *link = node;
